endpoint for getting snapshot files with chunks + bugfixes
- fixed bug where it could only chunk CoreCount+10 files
- fixed bug which prevented snapshots which had files with no chunks
  from creating
- snapshot files store relative paths instead of full paths

+++ internal/server/dbm/models.go
@@ -30,7 +30,7 @@ type File struct {
 type FileChunk struct {
 	FileHash   []byte
 	ChunkHash  []byte
-	ChunkOrder int16
+	ChunkOrder int32
 }
 
 type Permission struct {
